@@ -18,7 +18,7 @@ description: example of dio log
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -26,10 +26,12 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  dio: ^3.0.0
+  cupertino_icons: ^1.0.1
+  dio: ^5.2.0
+
   dio_log:
     path: ../../dio_log
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
