@@ -1,6 +1,7 @@
+///需要的请求错误数据类
 class ErrOptions {
-  int id;
-  String errorMsg;
+  int? id;
+  String? errorMsg;
   ErrOptions({
     this.id,
     this.errorMsg,
