@@ -6,19 +6,21 @@ import 'dio_log.dart';
 /// Created by rich on 2019-07-31
 ///
 
-OverlayEntry itemEntry;
+OverlayEntry? itemEntry;
 
 ///显示全局悬浮调试按钮
-showDebugBtn(BuildContext context, {Widget button, bool isDelay = true}) {
+showDebugBtn(BuildContext context, {Widget? button, Color? btnColor}) async {
   ///widget第一次渲染完成
-  WidgetsBinding.instance.addPostFrameCallback((_) {
+  try {
+    await Future.delayed(Duration(milliseconds: 500));
     dismissDebugBtn();
     itemEntry = OverlayEntry(
-        builder: (BuildContext context) => button ?? DraggableButtonWidget());
+        builder: (BuildContext context) =>
+            button ?? DraggableButtonWidget(btnColor: btnColor));
 
     ///显示悬浮menu
-    Overlay.of(context)?.insert(itemEntry);
-  });
+    Overlay.of(context)?.insert(itemEntry!);
+  } catch (e) {}
 }
 
 ///关闭悬浮按钮
@@ -34,13 +36,15 @@ bool debugBtnIsShow() {
 
 class DraggableButtonWidget extends StatefulWidget {
   final String title;
-  final Function onTap;
+  final Function? onTap;
   final double btnSize;
+  final Color? btnColor;
 
   DraggableButtonWidget({
     this.title = 'http log',
     this.onTap,
     this.btnSize = 66,
+    this.btnColor,
   });
 
   @override
@@ -50,8 +54,8 @@ class DraggableButtonWidget extends StatefulWidget {
 class _DraggableButtonWidgetState extends State<DraggableButtonWidget> {
   double left = 30;
   double top = 100;
-  double screenWidth;
-  double screenHeight;
+  late double screenWidth;
+  late double screenHeight;
 
   @override
   void initState() {
@@ -72,10 +76,10 @@ class _DraggableButtonWidgetState extends State<DraggableButtonWidget> {
       );
     };
     Widget w;
-    Color primaryColor = Theme.of(context).primaryColor;
+    Color primaryColor = widget.btnColor ?? Theme.of(context).primaryColor;
     primaryColor = primaryColor.withOpacity(0.6);
     w = GestureDetector(
-      onTap: widget.onTap ?? tap,
+      onTap: widget.onTap as void Function()? ?? tap,
       onPanUpdate: _dragUpdate,
       child: Container(
         width: widget.btnSize,
