@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-copyClipboard(BuildContext context, String value) {
-  var snackBar = SnackBar(content: Text('copy $value to clipboard'));
-  Scaffold.of(context).showSnackBar(snackBar);
-  Clipboard.setData(ClipboardData(text: value));
+///复制到粘贴板
+copyClipboard(BuildContext context, String? value) {
+  var snackBar =
+      SnackBar(content: Text('$value\n\n copy success to clipboard'));
+  ScaffoldMessenger.of(context).showSnackBar(snackBar);
+  Clipboard.setData(ClipboardData(text: value ?? 'null'));
 }
