@@ -1,7 +1,7 @@
 import 'dart:convert';
 
+import 'package:dio_log/utils/copy_clipboard.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
 ///
 /// Created by rich on 2019-07-17
@@ -12,7 +12,7 @@ class JsonView extends StatefulWidget {
   final dynamic json;
 
   ///是否展开全部json
-  final bool isShowAll;
+  final bool? isShowAll;
 
   final double fontSize;
   JsonView({
@@ -26,7 +26,7 @@ class JsonView extends StatefulWidget {
 }
 
 class _JsonViewState extends State<JsonView> {
-  Map<String, bool> showMap = Map();
+  Map<String, bool?> showMap = Map();
 
   ///当前节点编号
   int currentIndex = 0;
@@ -47,7 +47,7 @@ class _JsonViewState extends State<JsonView> {
     if (type == JsonType.object) {
       w = _buildObject(widget.json);
     } else if (type == JsonType.array) {
-      List list = widget.json as List;
+      List? list = widget.json as List?;
       w = _buildArray(list, '');
     } else {
       var je = JsonEncoder.withIndent('  ');
@@ -58,7 +58,7 @@ class _JsonViewState extends State<JsonView> {
   }
 
   ///构建object节点的展示
-  Widget _buildObject(Map<String, dynamic> json, {String key}) {
+  Widget _buildObject(Map<String, dynamic>? json, {String? key}) {
     List<Widget> listW = [];
 
     ///增加一个节点
@@ -76,7 +76,7 @@ class _JsonViewState extends State<JsonView> {
     ///object展示内容
     if (_isShow(currentIndex)) {
       List<Widget> listObj = [];
-      json.forEach((k, v) {
+      json!.forEach((k, v) {
         Widget w;
         JsonType type = getType(v);
         if (type == JsonType.object) {
@@ -110,7 +110,7 @@ class _JsonViewState extends State<JsonView> {
   }
 
   ///构建array节点的展示
-  Widget _buildArray(List listJ, String key) {
+  Widget _buildArray(List? listJ, String key) {
     List<Widget> listW = [];
 
     ///增加一个节点
@@ -128,11 +128,17 @@ class _JsonViewState extends State<JsonView> {
 
     ///添加key的点击事件
     ///添加key的展示
-    listW.add(_wrapFlex(currentIndex, keyW));
+    listW.add(GestureDetector(
+      behavior: HitTestBehavior.translucent,
+      onLongPress: () {
+        _copy(listJ.toString());
+      },
+      child: _wrapFlex(currentIndex, keyW),
+    ));
 
     if (_isShow(currentIndex)) {
       List<Widget> listArr = [];
-      listJ.forEach((val) {
+      listJ!.forEach((val) {
         var type = getType(val);
         Widget w;
         if (type == JsonType.object) {
@@ -207,7 +213,7 @@ class _JsonViewState extends State<JsonView> {
   bool _isShow(int key) {
     ///说明是根节点
     if (key == 1) return true;
-    if (widget.isShowAll) {
+    if (widget.isShowAll!) {
       return showMap[key.toString()] ?? true;
     } else {
       return showMap[key.toString()] ?? false;
@@ -217,12 +223,12 @@ class _JsonViewState extends State<JsonView> {
   ///展开合上的切换
   _flexSwitch(String key) {
     showMap.putIfAbsent(key, () => false);
-    showMap[key] = !showMap[key];
+    showMap[key] = !showMap[key]!;
     setState(() {});
   }
 
   ///展开合上所有
-  _flexAll(bool flex) {
+  _flexAll(bool? flex) {
     showMap.forEach((k, v) {
       showMap[k] = flex;
     });
@@ -247,11 +253,9 @@ class _JsonViewState extends State<JsonView> {
     );
   }
 
+  ///复制到手机粘贴板
   _copy(value) {
-    var snackBar =
-        SnackBar(content: Text('"$value"\n\n copy success to clipboard'));
-    Scaffold.of(context).showSnackBar(snackBar);
-    Clipboard.setData(ClipboardData(text: value?.toString()));
+    copyClipboard(context, value);
   }
 }
 
