@@ -1,6 +1,5 @@
 import 'package:dio_log/bean/net_options.dart';
 import 'package:dio_log/dio_log.dart';
-import 'package:dio_log/widget/json_view.dart';
 import 'package:flutter/material.dart';
 
 class LogResponseWidget extends StatefulWidget {
@@ -26,18 +25,13 @@ class _LogResponseWidgetState extends State<LogResponseWidget>
       children: <Widget>[
         Row(
           children: <Widget>[
-            RaisedButton(
-              onPressed: () {
-                copyClipboard(context, toJson(json));
-              },
-              child: Text('copy json'),
-            ),
             SizedBox(width: 10),
             Text(isShowAll ? 'shrink all' : 'expand all'),
-            Checkbox(
+            Switch(
               value: isShowAll,
               onChanged: (check) {
                 isShowAll = check;
+
                 setState(() {});
               },
             ),
@@ -61,13 +55,36 @@ class _LogResponseWidgetState extends State<LogResponseWidget>
             color: Colors.red,
           ),
         ),
+        _buildJsonView('headers:', response?.headers),
+        _buildJsonView('response.data:', json),
+      ],
+    ));
+  }
+
+  ///构建json树的展示
+  Widget _buildJsonView(key, json) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: <Widget>[
+        ElevatedButton(
+          onPressed: () {
+            copyClipboard(context, toJson(json));
+          },
+          child: Text('copy json'),
+        ),
+        Text(
+          '$key',
+          style: TextStyle(
+            fontSize: fontSize,
+          ),
+        ),
         JsonView(
           json: json,
           isShowAll: isShowAll,
           fontSize: fontSize,
         ),
       ],
-    ));
+    );
   }
 
   @override
