@@ -1,12 +1,14 @@
 import 'dart:convert';
 
+///返回json格式的String
 toJson(dynamic data) {
   var je = JsonEncoder.withIndent('  ');
   var json = je.convert(data);
   return json;
 }
 
-String map2Json(Map map) {
+///返回json格式的String
+String map2Json(Map? map) {
   if (map == null) {
     return '';
   }
