@@ -1,7 +1,7 @@
+///拼接实际请求链接
 String getRealUrl(String url, Map<String, dynamic> params) {
-  ///拼接实际请求链接
   StringBuffer str = StringBuffer(url);
-  if (params?.length ?? 0 > 0) {
+  if (params.length as bool? ?? 0 > 0) {
     str.write('?');
     var hasAnd = false;
     params.forEach((key, value) {
